requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib
python-multipart
torch --index-url https://download.pytorch.org/whl/cpu
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
pyjwt
pydantic

faiss-cpu
sentence-transformers
pypdf
pandas